Created Base template and added other webpages
Created a base template and updated the index template to match the base
template. I also created the anime, manga, community, industry, watch,
and help webpages that are using the base template.  I also created
working links to the newly created webpages so some form of navigation
on the website is live.